Mer gracefull nedskalering av scope (#76)
* forbedret lesbarhet

* fjernet kommentar

* erstattet gammel kode for å slette alle tokens

* setter default-verdi eksplisitt

* fjernet print fra test..

* rettet skrivefeil i accessiblitylabel

* satt button-trait på list-items

* UU skjuler element som kun brukes på iPad

* UU forbedrer labels i loginview

* setter trait-header på "MINE MAPPER" i mappeliste

* setter UU-label på strekkode

* Åpner ekstern lenke i SFSafariViewController

* endret feil default-verdi

* bruker feilkode ved manglende oauth-token

* fjerner utkommert kode for tidligere caching

* logger ut om token ikke finnes lokalt

* fjerner kode for refresh ved flere tokens lokalt

* logger ut bruker uten lokal-token på kontoview

* bruker riktigere feilmelding

* bruker felles utlogging/vis-login

* retter på feilmelding

* fjernet revoke av gcm tokens ved utlogging etter LA-feil

* fjernet logging

* oppdatert versjonsnummer

* fjernet ubrukt metode

* sletter kun om token eksisterer

* fjernet debug-prints

+++ Digipost/OAuthToken.swift
@@ -125,20 +125,6 @@ struct AuthenticationLevel {
         saveToken()
     }
     
-    func printContent(_ title: String){
-        print("auth ****************** printContent *********************");
-        print("auth : \(title)")
-        if let token = LUKeychainAccess.standard().object(forKey: kOAuth2Token) as? OAuthToken {
-            
-            print(token)
-            print("auth scope:\(token.scope)")
-            print("auth access \(token.accessToken)")
-            print("auth refresh \(token.refreshToken)")
-            print("auth scope \(token.scope)")
-        }
-        print("auth **************** printContent end ***********************");
-    }
-    
     // bug in swift compiler requires to set all instance variables before returning nil from an initializer
     fileprivate func setAllInstanceVariablesToNil() {
         self.refreshToken = nil
@@ -155,7 +141,9 @@ struct AuthenticationLevel {
     }
     
     @objc class func removeToken() {
-        LUKeychainAccess.standard().deleteObject(forKey: kOAuth2Token)
+        if self.getToken() != nil {
+            LUKeychainAccess.standard().deleteObject(forKey: kOAuth2Token)
+        }
     }
     
     @objc class func isUserLoggedIn() -> Bool {
@@ -239,13 +227,6 @@ struct AuthenticationLevel {
         }
         return true
     }
-
-    @objc func removeFromKeychainIfNoAccessToken() {
-        if accessToken == nil {
-            OAuthToken.removeToken()
-        }
-    }
-    
     
     @objc func setAccessTokenAndScope(_ accessToken: NSString, scope: NSString) {
         self.accessToken = accessToken as String
